+++ cmd/server/main.go
@@ -1,25 +1,11 @@
 package main
 
 import (
-	"flag"
 	"log"
 	"net/http"
 
 	"github.com/go-chi/chi"
 )
-
-// переменная flagRunAddr содержит адрес и порт для запуска сервера
-var flagRunAddr string
-
-// parseFlags обрабатывает аргументы командной строки
-// и сохраняет их значения в соответствующих переменных
-func parseFlags() {
-	// регистрируем переменную flagRunAddr
-	// как аргумент -a со значением localhost:8080 по умолчанию
-	flag.StringVar(&flagRunAddr, "a", "localhost:8080", "address and port to run server")
-	// парсим переданные серверу аргументы в зарегистрированные переменные
-	flag.Parse()
-}
 
 // main — точка входа приложения.
 // Инициализирует роутер chi, создаёт хранилище метрик и настраивает маршруты:
